Split code into multiple classes + transform to a Python package
