+++ setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/xacro_live
+[install]
+install-scripts=$base/lib/xacro_live
